move tracer into separate package
